% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bearing.R
\name{add_bearings}
\alias{add_bearings}
\title{Add track bearing}
\usage{
add_bearings(stations)
}
\arguments{
\item{stations}{Dataframe (tibble) of the samples stations including the attributes
"Lon_start_DD", "Lat_start_DD", "Lon_stop_DD", "Lat_stop_DD".}
}
\value{
Returns the input dataframe (tibble) with an added attributes called
"Bearing_calc" and "Bearing_calc". Stations without start and stop coordinates have
NA in these added columns.
}
\description{
Function to add the initial bearing based on the start and stop coordinates
of the track.
}
\references{
geosphere R-package
}
