% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_track_length.R
\name{add_track_length}
\alias{add_track_length}
\alias{add_track_length_GPS}
\alias{add_track_length_Odometer}
\title{Add calculated track length}
\usage{
add_track_length_GPS(stations)

add_track_length_Odometer(stations)
}
\arguments{
\item{stations}{(required) Dataframe (tibble) with all stations and the attributes
(columns) "Lon_start_DD", "Lat_start_DD","Lon_stop_DD", and "Lat_stop_DD".}
}
\value{
This function returns the input tibble with an added column called
"Track_dist_m_GPS".
}
\description{
The functions \code{add_track_length_GPS} and \code{add_track_length_Odometer}
add a column with tracklength based on track start and stop GPS coordinates and
on Odometer ticks respectively.
}
\references{
\itemize{
\item{geosphere R-package}
}
}
