% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_track_length.R
\name{add_track_length_Odometer}
\alias{add_track_length_Odometer}
\title{Add track length based on the odometer}
\usage{
add_track_length_Odometer(stations)
}
\arguments{
\item{stations}{(required) Dataframe (tibble) with all stations and the attributes
(columns) "Lon_start_DD", "Lat_start_DD","Lon_stop_DD", and "Lat_stop_DD".}
}
\value{
This function returns the input tibble with an added column called
"Track_dist_m_BB" and "Track_dist_m_SB".
}
\description{
The function adds a column with track length in meters based on odometer ticks (one tick is two meters).
}
\references{
\itemize{
\item{geosphere R-package}
}
}
