% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_track_midpoints.R
\name{add_track_midpoints}
\alias{add_track_midpoints}
\title{Add track midpoints}
\usage{
add_track_midpoints(stations)
}
\arguments{
\item{stations}{Dataframe (tibble) of the samples stations including the attributes
"Lon_start_DD", "Lat_start_DD", "Lon_stop_DD", "Lat_stop_DD".}
}
\value{
Returns the input dataframe (tibble) with an added attributes called
"Lon_DD_calc" and "Lat_DD_calc". Stations without start and stop coordinates have
NA in these added columsn.
}
\description{
Function to add the found geographic midpoints of each sample track to the stations dataframe (tibble).
}
\references{
geosphere R-package
}
