% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{bathymetry}
\alias{bathymetry}
\title{TripleD NOAA bathymetry}
\format{Bathymetry data collected from NOAA with the marmap package.
Format is xyz data in tibble, x is longitude, y is latitude, z is altitude
(so negative values are depth).}
\usage{
bathymetry
}
\description{
Data frame (tibble) xyz bathymetry data collected from NOAA with the marmap
R-package based on the min and max coordinates in the database.
}
\references{
\itemize{
marmap R-package
}
}
\keyword{datasets}
