% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_water_depth.R
\name{collect_bathymetry}
\alias{collect_bathymetry}
\title{Collect NOAA bathymetry}
\usage{
collect_bathymetry(stations, buffer = 2, resolution = 1)
}
\arguments{
\item{stations}{Dataframe (tibble) with the attributes
"Lat_DD_midpt", "Lat_start_DD", "Lat_stop_DD", "Lon_DD_midpt", "Lon_start_DD", "Lon_stop_DD".}

\item{buffer}{The size of the buffer zone in decimal degrees around the
min and max coordinates in the dataframe. Default = 2.}

\item{resolution}{The resolution in minutes the bathymetry is collected in.
A lower resolution provides better depth estimates from bathymetry, but takes
a longer time to collect from the NOAA database. Default resolution is 1 minute.}
}
\value{
Returns a xyz dataframe (x = lon, y = lat, z = altitude i.e. negative values are depth)
with bathymetry of a resolution of 1 minute.
}
\description{
This function collects NOAA bathymetry with the marmap R-pacakge
based on the min and max coordinates in the database.
}
\references{
marmap R-package
}
