% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_complete_database.R
\name{collect_bioconversion_WORMS}
\alias{collect_bioconversion_WORMS}
\title{Collect species names in bioconversion file from WoRMS}
\usage{
collect_bioconversion_WORMS(
  conversion_data = NULL,
  input_folder = "inputfiles",
  out_folder = "data",
  as_CSV = FALSE
)
}
\arguments{
\item{conversion_data}{Dataframe with bioconversion data matching the given requirements from the
attributes_bioconversion file. If NULL (default) the bioconversion.csv will be searched for and
loaded from the input_folder.}

\item{input_folder}{The folder where to find the bioconversion.csv file. Default is 'inputfiles'.}

\item{out_folder}{The external data is stored in this folder. Default is 'data'.}

\item{as_CSV}{If you also want to store the collected external data as CSV, set to TRUE. Default is FALSE.}
}
\value{
This function does not return an object, but stores the information in the specified
\code{out_folder} under the names 'worms_conversion.rda' and 'conversion_data.rda'.
}
\description{
This function queries reported species names from the bioconversion input file against
WoRMS and calculates average conversion factors and regression formula's for
higher taxonomic levels.
}
\details{
Taxonomic data is collected from the WoRMS database using the \code{worrms} R-package.
You need internet connection to do this.
The reported taxonomic names of the specimens in the bioconversion.csv file
are matched against the WoRMS database (also fuzzy matches, i.e. where typos and phonetic spelling is allowed).
}
