% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_worms_taxonomy.R
\name{collect_from_worms}
\alias{collect_from_worms}
\alias{get_worms_taxonomy}
\title{Add WoRMs taxonomy}
\usage{
collect_from_worms(species_names, fuzzy = FALSE)

get_worms_taxonomy(species_names)
}
\arguments{
\item{species_names}{Only the species names.}

\item{fuzzy}{Uses the TAXAMATCH algorithm to match species names.}

\item{species}{Tibble with reported species names.}
}
\description{
These functions are used to check the reported species names against the WoRMs datbase.
}
\details{
Searches in batches of 50.
}
\references{
worrms R-package
}
