% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_complete_database.R
\name{collect_species_WORMS}
\alias{collect_species_WORMS}
\title{Collect reported species names from WoRMS}
\usage{
collect_species_WORMS(
  species = NULL,
  data_folder = "data",
  out_folder = "data",
  as_CSV = FALSE
)
}
\arguments{
\item{species}{The initial database for species with reported specimen names.
If NULL (default) the function will automatically search the data_folder for 'species_initial.rda'.}

\item{data_folder}{If the species database is not provided, the function will search
for it (species_initial.rda') in this folder. Default is 'data'.}

\item{out_folder}{The external data is stored in this folder. Default is 'data'.}

\item{as_CSV}{If you also want to store the collected external data as CSV, set to TRUE. Default is FALSE.}
}
\value{
This function does not return an object, but stores the information in the specified
\code{out_folder} under the name 'worms.rda'.
}
\description{
This function collects accepted species names from WoRMS for all reported species.
}
\details{
Taxonomic data is collected from the WoRMS database using the \code{worrms} R-package.
You need internet connection to do this.
The reported taxonomic names of the specimens in the initial database
is matched against the WoRMS database (also fuzzy matches, i.e. where typos and phonetic spelling is allowed).
}
