% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{community}
\alias{community}
\alias{community_matrix_counts}
\alias{community_matrix_biomass}
\title{TripleD community data}
\format{A dataframe with the following columns:
\describe{
\item{\code{File}}{File name of the csv that contained the data entry.}
\item{\code{Station}}{A unique sampling ID. Usually the cruiseID combined with a number. Multiple sampling stations can occur at the same geographic location (i.e. have the same station name).}
\item{\code{Unit}}{}
\item{\code{Factor}}{}
\item{\code{Species}}{Scientific species name of the sampled individual.}
\item{\code{Common_name}}{Common name of the sampled individual.}
\item{\code{Count}}{The total count of individuals}
\item{\code{Size}}{The mean size of the species reported in the unit as in column "Unit".}
\item{\code{Size_sd}}{The standard deviation in size}
\item{\code{WetWeight}}{The total wet weight of the species reported in grams.}
}}
\usage{
community

community_matrix_counts

community_matrix_biomass
}
\description{
Data frame (tibble) with the number of individuals and total wet weight per species
per sampling station. Also in matrix form to use in multivariate statistics.
}
\details{
The data frame 'community' is in tidy format (all data in columns).
The data frames 'community_matrix_counts' and 'community_matrix_biomass' are in the
form of community matrices (sample stations as rows, species as columns), and can be used
in multivariate statistics.
}
\references{
\itemize{
\item{Rob Witbaard, NICO 10, 64PE438}
}
}
\keyword{datasets}
