% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_complete_database.R
\name{complete_database}
\alias{complete_database}
\title{Complete database with external data and calculations}
\usage{
complete_database(data_folder = "data", out_folder = "data", bathymetry = NULL)
}
\arguments{
\item{data_folder}{This is the folder that contains the initial database.
Default is 'data', as created by the \code{construct_database} function.}

\item{out_folder}{This is the folder where you want your improved, but not yet finalized database,
to be stored. Default is 'data'.}

\item{bathymetry}{(optional) Water depth can be added to the database if you have a
bathymetry file. If no such file is given, bathymetry is collected from NOAA.}
}
\value{
This function does not return an object, but stores 'stations_additions.rda'
and 'species_additions.rda' to the out_folder.
}
\description{
This function contains a workflow to add information to the original data
through external datasets and calculations. Please review this workflow as
explained in \code{Details}. You can also create your own workflow, for example by using
other functions contained in this package.
}
\details{
The initial database created with \code{construct_database} only contains
raw data as imported from the CSV files. Sometimes data is missing (e.g.
water depth is optional and not always reported) or you wish to include
extra information (e.g. taxonomic information). This workflow is a chain
of functions to improve the database with external data and calculations.
The chain of functions is explained here, but you are of course free to
adjust this chain to your own wishes.
\cr
#' The latest accepted name for the specimen is added to the database including the taxonomic classification (phylum, class, order, family, genus).
}
\seealso{
\code{add_track_midpoints}, \code{add_track_length_GPS},
\code{add_track_length_Odometer}, \code{add_water_depth}.
}
