% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_worms_taxonomy.R
\name{connect_to_worms}
\alias{connect_to_worms}
\title{Connect to WoRMS API}
\usage{
connect_to_worms(species_names, fuzzy = FALSE)
}
\arguments{
\item{species_names}{Character vector with taxa names.}

\item{fuzzy}{Uses the TAXAMATCH algorithm to fuzzy match species names.}
}
\description{
This function collects WoRMS entries from their database using the reported taxa in batches of 50,
using either the regular function or fuzzy function (taxamatch).
}
\details{
Searches in batches of 50.
}
\references{
worrms R-package
}
