% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_construct_database.R
\name{construct_database}
\alias{construct_database}
\title{Construct initial database}
\usage{
construct_database(
  in_folder = "inputfiles",
  out_folder = "data",
  as_CSV = FALSE
)
}
\arguments{
\item{in_folder}{This folder must contain two subfolders called 'Species' and 'Stations' that contain
properly formatted CSV files. These CSV files can be requested from the NIOZ Data Archiving System.
Default is "inputfiles".}

\item{out_folder}{The function will create a new directoy to store the newly created data.
Default is "data".}
}
\value{
This function does not return any objects. It will throw errors if anything is wrong with
your initial data. If there are no errors, it will store Rdata in the newly created folder 'out_folder'.
}
\description{
This function loads CSV files, checks the data format, and constructs an initial table
with only raw data.
}
\details{
Make sure this function is executed in the same working directory that contains the in_folder directory.
This function checks the data format of the CSV files contained by the 'Species' and 'Stations' folder
contained by the 'in-folder'. If the functions finds any irregularities, an error will be thrown.
}
