% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_community_matrix.R
\name{get_community_matrix}
\alias{get_community_matrix}
\title{Get community matrix from the TripleD database}
\usage{
get_community_matrix(
  database,
  taxonomic_level,
  abundance_measure,
  include_incomplete = TRUE
)
}
\arguments{
\item{database}{A dataframe, database output from the TripleD package.}

\item{taxonomic_level}{The taxonomic level for which you want to create
a community matrix. Options: \itemize{
\item{all}
\item{species (default)}
\item{genus}
\item{family}
\item{order}
\item{class}
\item{phylum}
}}

\item{abundance_measure}{The value in the cells. Options: \describe{
\item{presence_absence}{only 1 (present) or 0 (absence)}
\item{Density_nr_per_m2}{individual count per square meter}
\item{Biomass_g_per_m2"}{ash-free dry weight in grams per square meter}
\item{Density_nr_per_m3}{individual count per cubic meter}
\item{Biomass_g_per_m3}{ash-free dry weight in grams per cubic meter}
}}

\item{include_incomplete}{Include values based on incomplete data
i.e. underestimations? (Default TRUE)}
}
\value{
Returns a dataframe with StationID as first row.
}
\description{
This functions lets you extract a community matrix
(i.e. stations in rows, taxa in columns) from the TripleD databas.
}
\details{
Missing data filled with 0.
}
\examples{
get_community_matrix(database, "species", "Biomass_g_per_m2")
get_community_matrix(database, "phylum", "Density_nr_per_m2", include_incomplete = FALSE)
}
