% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_worms_taxonomy.R
\name{get_valid_names}
\alias{get_valid_names}
\title{Get dataframe with valid taxa names}
\usage{
get_valid_names(species_names)
}
\arguments{
\item{species_names}{Character vector with taxa names.}
}
\value{
Returns a dataframe.
}
\description{
This function checks all reported taxa against the WoRMS databse, and
returns a dataframe with currently accepted valid species names.
}
\details{
This function collects WoRMS entries based on the given species names, and reports the valid name
and rank that is found. It first finds all direct matches, and then does a fuzzy match for queries that do
not have a direct match against the database.
Only extant species are considered, and if multiple matches are returned, either
1) use the only accepted name,
2) if all are unaccepted, use the only valid name, or
3) use an empty record if there are multiple valid names returned.
}
\references{
worrms R-package
}
