% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_to_weight.R
\name{length_to_weight}
\alias{length_to_weight}
\title{Calculate weight from length}
\usage{
length_to_weight(length, A, B)
}
\arguments{
\item{length}{Length of organism in mm.}

\item{A}{The A factor in the power function A * length^B.}

\item{B}{The B exponent in the power function A * length^B.}
}
\value{
Returns the weight as Ash Free Dry Weight (AFDW) in grams (g).
}
\description{
This function calculates Ash Free Dry Weight (g) from length (mm) using the power function
weight = A * length^B.
with the coefficients taken from regression lines.
}
