% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconversion.R
\name{prepare_bioconversion}
\alias{prepare_bioconversion}
\title{Prepare the bioconversion file}
\usage{
prepare_bioconversion(
  conversion_data = NULL,
  worms = NULL,
  input_folder = "inputfiles",
  data_folder = "data",
  out_folder = "data",
  as_CSV = FALSE
)
}
\arguments{
\item{conversion_data}{Dataframe with bioconversion data matching the given requirements from the
attributes_bioconversion file. If NULL (default) the bioconversion.csv will be searched for and
loaded from the input_folder.}

\item{worms}{Dataframe with query and valid names from worms. If NULL (default), the file
"worms.rda" is searched for in the data_folder.}

\item{input_folder}{The folder where to find the bioconversion.csv file. Default is 'inputfiles'.}

\item{out_folder}{The external data is stored in this folder. Default is 'data'.}

\item{as_CSV}{If you also want to store the collected external data as CSV, set to TRUE. Default is FALSE.}
}
\value{
This function does not return an object, but stores a dataframe in the specified
\code{out_folder} under the name 'conversion_data.rda'.
}
\description{
This wrapper function adds valid taxon names to the bioconversion file and also calculates
mean conversion values for all higher taxonomic levels.
}
