% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{sampling_stations}
\alias{sampling_stations}
\title{TripleD sampling stations}
\format{A dataframe with the following columns:
\describe{
\item{\code{File}}{File name of the csv that contained the data entry.}
\item{\code{Station}}{A unique sampling ID. Usually the cruiseID combined with a number. Multiple sampling stations can occur at the same geographic location (i.e. have the same station name).}
\item{\code{Station_name}}{A generic name of the station that may contain some information on the location.}
\item{\code{Year}}{The year when the specific station was sampled.}
\item{\code{Month}}{The month when the specific station was sampled (1 to 12, Jan to Dec).}
\item{\code{Day}}{The day of the month when the specific station was sampled (1 to 31).}
\item{\code{Time}}{The time when the specific station was sampled.}
\item{\code{Lat}}{The average latitude (GPS coordinates) of the sampling station (in between the start and end position of the ship during the dredge) expressed in decimal degrees. Reference coordinate system is WGS84.}
\item{\code{Lon}}{The average longitude (GPS coordinates) of the sampling station (in between the start and end position of the ship during the dredge) expressed in decimal degrees. Reference coordinate system is WGS84.}
\item{\code{Distance}}{Distance dredged at this sampling location expressed in meters.}
\item{\code{Direction}}{Average bearing of the ship during the dredge, expressed in ??.}
\item{\code{Blade_width}}{The width of the used blade expressed in cm.}
\item{\code{Blade_depth}}{The depth of the used blade expressed in cm.}
}}
\usage{
sampling_stations
}
\description{
Data frame (tibble) with information on all stations where samples with the
TripleD were collected.
}
\details{
Details here.
}
\references{
\itemize{
\item{Rob Witbaard, NICO 10, 64PE438}
}
}
\keyword{datasets}
