% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{species}
\alias{species}
\title{TripleD species data}
\format{A dataframe with the following columns:
\describe{
\item{\code{File}}{File name of the csv that contained the data entry.}
\item{\code{Station}}{A unique sampling ID. Usually the cruiseID combined with a number. Multiple sampling stations can occur at the same geographic location (i.e. have the same station name).}
\item{\code{Unit}}{}
\item{\code{Factor}}{}
\item{\code{Species}}{Scientific species name of the sampled individual.}
\item{\code{Common_name}}{Common name of the sampled individual.}
\item{\code{Count}}{How many individuals are included in this data entry? Should be 1 if the reported information is per individual.}
\item{\code{Size}}{The size of the individual reported by the unit reported in the column "Unit".}
\item{\code{WetWeight}}{The wet weight of the (group of) individuals reported in grams. If sizes are reported for individuals, but wet weight is only known for the collection of individuals per species, then report a wet weight of zero, except for one individual. In the analysis the total wet weight per species is calculated.}
}}
\usage{
species
}
\description{
Data frame (tibble) with count and sometimes measurements of all sampled individuals.
Raw biological data.
Preferably each data entry (row) concerns an individual. However, often a small group
of individuals from the same species was measured and the given metrics concern the
group as a whole.
}
\details{
Researchers collected species data during TripleD field studies.
They are encourage to store this raw data in the NIOZ Data Archiving System
(DAS) so that other users of the database can retrace the origin of the data.
The raw data was used to construct CSV files in the right database format.
A description of this process should be provided by the contributers to the
database in the vignette "cleaning_data.Rmd". The CSV files' format was
checked with the code in "construct_database.R" and merged into a datframe
(tibble from the dplyr package) called "species". This dataframe contains
only original data, so no data from other sources or calculations based on
assumptions.
}
\references{
\itemize{
\item{Rob Witbaard (NIOZ), NICO 10, 64PE438}
\item{Marc Lavaleye (NIOZ), Access database}
}
}
\keyword{datasets}
