% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{stations}
\alias{stations}
\alias{stations_additions}
\title{TripleD stations}
\format{A dataframe with the following columns:
\describe{
\item{\code{File}}{File name of the csv that contained the data entry.}
\item{\code{Station}}{A unique sampling ID. Usually the cruiseID combined with a number. Multiple sampling stations can occur at the same geographic location (i.e. have the same station name).}
\item{\code{Station_name}}{A generic name of the station that may contain some information on the location.}
\item{\code{Year}}{The year when the specific station was sampled.}
\item{\code{Month}}{The month when the specific station was sampled (1 to 12, Jan to Dec).}
\item{\code{Day}}{The day of the month when the specific station was sampled (1 to 31).}
\item{\code{Time}}{The time when the specific station was sampled.}
\item{\code{Lat}}{The average latitude (GPS coordinates) of the sampling station (in between the start and end position of the ship during the dredge) expressed in decimal degrees. Reference coordinate system is WGS84.}
\item{\code{Lon}}{The average longitude (GPS coordinates) of the sampling station (in between the start and end position of the ship during the dredge) expressed in decimal degrees. Reference coordinate system is WGS84.}
\item{\code{Distance}}{Distance dredged at this sampling location expressed in meters.}
\item{\code{Direction}}{Average bearing of the ship during the dredge, expressed in ??.}
\item{\code{Blade_width}}{The width of the used blade expressed in cm.}
\item{\code{Blade_depth}}{The depth of the used blade expressed in cm.}
}}
\usage{
stations

stations_additions
}
\description{
Data frame (tibble) with information on all stations where samples with the
TripleD were collected.
}
\details{
Researchers collected raw data during TripleD field studies about the
sampling stations. They are encouraged to upload this raw data to the NIOZ
data archiving system (DAS) so that other database users can retrace the origin
of the data. CSV files in the correct database format were created by
researchers based on their original raw data.
A description of how the csv files were created based on this raw data
can be found in the vignette "cleaning_data.Rmd". The CSV files' format
was checked with the code in "construct_database.R" and merged into the
dataframe (tibble format from the dplyr package) called "stations".
Therefore, this dataframe should only include original data and no data from
other sources or calculations based on assumptions.
}
\references{
\itemize{
\item{Rob Witbaard (NIOZ), NICO 10, 64PE438}
\item{Marc Lavaleye (NIOZ), Access database}
}
}
\keyword{datasets}
