% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{worms}
\alias{worms}
\title{TripleD WoRMS taxonomy}
\format{Tibble with queried name, its valid name according to WoRMS, its taxonomic rank,
taxonomic classification, the match type (exact, nr of differences, phonetic, etc.), if the match
was found with fuzzy searching (isFuzzy) or wether there was no match at all (hasNoMatch), and
other information automatically generated by WoRMS.}
\usage{
worms
}
\description{
Data frame (tibble) with taxonomy data of the reported species in the TripleD dataset
taken from the World Register of Marine Species database (package worrms). Added columns are
'isFuzzy' and 'hasNoMatch', both boolean variables.
}
\references{
\itemize{
worrms R-package
}
}
\keyword{datasets}
