% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_water_depth.R
\name{add_water_depth}
\alias{add_water_depth}
\title{Add water depth}
\usage{
add_water_depth(stations, bathymetry = NULL)
}
\arguments{
\item{stations}{Dataframe (tibble) with the attributes
"Lat_DD", "Lat_start_DD", "Lat_stop_DD", "Lon_DD", "Lon_start_DD", "Lon_stop_DD".}

\item{bathymetry}{An xyz dataframe (tibble) or object of class bathy created by
the marmap package. Default = NULL, uses \code{collect_bathymetry()} function to first download data.}
}
\value{
Returns a xyz dataframe (x = lon, y = lat, z = altitude i.e. negative values are depth)
with bathymetry of a resolution of 1 minute.
}
\description{
This function adds an attribute with water depth based on track midpoints and the
collected NOAA bathymetry with the marmap R-pacakge using the function \code{get.depth()}.
}
\references{
maramp R-package
}
