% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_water_depth.R
\name{collect_from_NOAA}
\alias{collect_from_NOAA}
\title{Collect bathymetry from NOAA}
\usage{
collect_from_NOAA(
  stations = NULL,
  buffer = 2,
  resolution = 1,
  data_folder = "data",
  out_folder = "data",
  lats = NULL,
  lons = NULL,
  as_CSV = FALSE
)
}
\arguments{
\item{stations}{The initial database for stations with track midpoints (attributes "Lat_DD_midpt" and "Lon_DD_midpt").
If NULL (default) the function will automatically search the data_folder for 'stations_initial.rda'.}

\item{buffer}{The size of the buffer zone in decimal degrees around the given
min and max coordinates. Default = 2.}

\item{resolution}{The resolution in minutes the bathymetry is collected in.
A lower resolution provides better depth estimates from bathymetry, but takes
a longer time to collect from the NOAA database. Default resolution is 1 minute.}

\item{data_folder}{If the stations database is not provided, the function will search
for it ('stations_initial.rda') in this folder. Default is 'data'.}

\item{out_folder}{The external data is stored in this folder. Default is 'data'.}

\item{lats}{(optional) You can specify latitudes you want to use to collect bathymetry:
a numeric vector with latitudes of which the min and max are found and used.
If not specified, the latitudes from the database are used.}

\item{lons}{(optional) You can specify longitudes you want to use to collect bathymetry:
a numeric vector with longitudes of which the min and max are found and used.
If not specified, the longitudes from the database are used.}

\item{as_CSV}{If you also want to store the collected external data as CSV, set to TRUE. Default is FALSE.}
}
\value{
This function does not return an object, but stores the information in the specified
\code{out_folder} under the names 'bathymetry.rda'.
}
\description{
This function collects NOAA bathymetry with the marmap R-pacakge
based on the min and max coordinates in the database.
}
\details{
Bathymetry data is collected from the NOAA database using the \code{marmap} R-package.
You need internet connection to do this. The default resolution is 1 minute.
You can also choose to collect your own bathymetry with a different resolution from NOAA by using the
package function \code{collect_bathymetry}, or not to collect bathymetry externally at all
if you have your own dataset.
}
\references{
marmap R-package
}
