% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_worms_taxonomy.R
\name{collect_from_WORMS}
\alias{collect_from_WORMS}
\title{Collect species names from WoRMS}
\usage{
collect_from_WORMS(
  species = NULL,
  conversion_data = NULL,
  input_folder = "inputfiles",
  data_folder = "data",
  out_folder = "data",
  as_CSV = FALSE
)
}
\arguments{
\item{species}{The initial database for species with reported specimen names.
If NULL (default) the function will automatically search the data_folder for 'species_initial.rda'.}

\item{conversion_data}{Dataframe with bioconversion data matching the given requirements from the
attributes_bioconversion file. If NULL (default) the bioconversion.csv will be searched for and
loaded from the input_folder.}

\item{input_folder}{The folder where to find the bioconversion.csv file. Default is 'inputfiles'.}

\item{data_folder}{If the species database is not provided, the function will search
for it (species_initial.rda') in this folder. Default is 'data'.}

\item{out_folder}{The external data is stored in this folder. Default is 'data'.}

\item{as_CSV}{If you also want to store the collected external data as CSV, set to TRUE. Default is FALSE.}
}
\value{
This function does not return an object, but stores the information in the specified
\code{out_folder} under the name 'worms.rda'.
}
\description{
This wrapper function collects accepted species names from WoRMS for all reported species in the
data files and the bioconversion file.
}
\details{
Taxonomic data is collected from the WoRMS database using the \code{worrms} R-package.
You need internet connection to do this.
All reported taxonomic names of the specimens in the initial database and in the bioconversion file
are matched against the WoRMS database (also fuzzy matches, i.e. where typos and phonetic spelling is allowed).
}
