% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{species}
\alias{species}
\title{TripleD species data}
\format{A dataframe with the following columns:
\describe{
\item{\code{File}}{File name of the csv that contained the data entry.}
\item{\code{Station}}{A unique sampling ID. Usually the cruiseID combined with a number. Multiple sampling stations can occur at the same geographic location (i.e. have the same station name).}
\item{\code{Unit}}{}
\item{\code{Factor}}{}
\item{\code{Species}}{Scientific species name of the sampled individual.}
\item{\code{Common_name}}{Common name of the sampled individual.}
\item{\code{Count}}{How many individuals are included in this data entry? Should be 1 if the reported information is per individual.}
\item{\code{Size}}{The size of the individual reported by the unit reported in the column "Unit".}
\item{\code{WetWeight}}{The wet weight of the (group of) individuals reported in grams. If sizes are reported for individuals, but wet weight is only known for the collection of individuals per species, then report a wet weight of zero, except for one individual. In the analysis the total wet weight per species is calculated.}
}}
\usage{
species
}
\description{
Data frame (tibble) with size and wet weight of all sampled individuals.
Raw biological data.
Preferably each data entry (row) concerns an individual. However, sometimes a small group
of individuals from the same species was measured and the given metrics concern the
group as a whole.
}
\details{
Details here.
}
\references{
\itemize{
\item{Rob Witbaard, NICO 10, 64PE438}
}
}
\keyword{datasets}
